backend/requirements.txt
# Core web framework
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic-settings
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
email-validator
